+++ include/ImageDownloader.hpp
@@ -9,12 +9,12 @@ struct MemoryBuffer {
     std::vector<unsigned char> data;
 };
 
-std::string getJson(const std::string &url, const long discordID);
+std::string getJson(const std::string &url, const long long discordID);
 std::string getIMG(const std::string IMGURL);
 
-std::string createUser(const std::string &url, const long discordID);
-std::string modifyUser(const std::string &url, const long discordID, std::string &actName, std::string &actDesc, std::string &actType);
-std::string setIMG(const std::string &url, const long discordID, const std::string &imgLoc);
+std::string createUser(const std::string &url, const long long discordID);
+std::string modifyUser(const std::string &url, const long long discordID, std::string &actName, std::string &actDesc, std::string &actType);
+std::string setIMG(const std::string &url, const long long discordID, const std::string &imgLoc);
 
 std::string getActName();
 std::string getActDesc();
